#8, Refactor MainActivity.
1. navigate activities through `Navigation`.
