+++ app/src/main/java/com/imeepwni/jetpack/data/JetPackFunction.kt
@@ -1,7 +1,13 @@
 package com.imeepwni.jetpack.data
 
+import androidx.annotation.StringRes
+import androidx.navigation.NavDirections
+
 /**
  * 作者：Created by guofeng on 2019/10/10
  * 邮箱：feng.guo@bees360.com
  */
-data class JetPackFunction(val name: String, val callback: () -> Unit)
+data class JetPackFunction(
+        @StringRes
+        val name: Int,
+        val action: NavDirections)
