+++ app/src/main/java/com/imeepwni/jetpack/main/MainFragment.kt
@@ -1,0 +1,51 @@
+package com.imeepwni.jetpack.main
+
+import android.os.Bundle
+import android.view.LayoutInflater
+import android.view.View
+import android.view.ViewGroup
+import androidx.recyclerview.widget.DividerItemDecoration
+import androidx.recyclerview.widget.LinearLayoutManager
+import androidx.recyclerview.widget.RecyclerView
+import com.imeepwni.jetpack.R
+import com.imeepwni.jetpack.app.BaseFragment
+import com.imeepwni.jetpack.data.JetPackFunction
+import com.imeepwni.jetpack.main.MainFragmentDirections.Companion.actionMainFragmentToDataBindingActivity
+import com.imeepwni.jetpack.main.MainFragmentDirections.Companion.actionMainFragmentToEncryptedFileActivity
+import com.imeepwni.jetpack.main.MainFragmentDirections.Companion.actionMainFragmentToLifecyclesActivity
+import com.imeepwni.jetpack.main.MainFragmentDirections.Companion.actionMainFragmentToNavigationActivity
+import kotlinx.android.synthetic.main.fragment_main.*
+
+/**
+ * 作者：Created by guofeng on 2019/11/21
+ * 邮箱：feng.guo@bees360.com
+ */
+class MainFragment : BaseFragment() {
+    private lateinit var jetPackFunctions: List<JetPackFunction>
+
+    override fun onCreateView(inflater: LayoutInflater, container: ViewGroup?, savedInstanceState: Bundle?): View? {
+        return inflater.inflate(R.layout.fragment_main, container, false)
+    }
+
+    override fun onViewCreated(view: View, savedInstanceState: Bundle?) {
+        super.onViewCreated(view, savedInstanceState)
+
+        initData()
+        initView()
+    }
+
+    private fun initData() {
+        jetPackFunctions = listOf(JetPackFunction(R.string.encrypted_file, actionMainFragmentToEncryptedFileActivity()),
+                JetPackFunction(R.string.data_binding, actionMainFragmentToDataBindingActivity()),
+                JetPackFunction(R.string.lifecycles, actionMainFragmentToLifecyclesActivity()),
+                JetPackFunction(R.string.navigation, actionMainFragmentToNavigationActivity()))
+    }
+
+    private fun initView() {
+        rv_content.run {
+            layoutManager = LinearLayoutManager(context, RecyclerView.VERTICAL, false)
+            adapter = FunctionAdapter(jetPackFunctions)
+            addItemDecoration(DividerItemDecoration(context, RecyclerView.VERTICAL))
+        }
+    }
+}
