+++ app/src/main/res/navigation/nav_main.xml
@@ -1,0 +1,48 @@
+<?xml version="1.0" encoding="utf-8"?>
+<navigation xmlns:android="http://schemas.android.com/apk/res/android"
+    xmlns:app="http://schemas.android.com/apk/res-auto"
+    xmlns:tools="http://schemas.android.com/tools"
+    android:id="@+id/nav_main"
+    app:startDestination="@id/mainFragment">
+
+    <fragment
+        android:id="@+id/mainFragment"
+        android:name="com.imeepwni.jetpack.main.MainFragment"
+        android:label="MainFragment"
+        tools:layout="@layout/fragment_main">
+        <action
+            android:id="@+id/action_mainFragment_to_dataBindingActivity"
+            app:destination="@id/dataBindingActivity" />
+        <action
+            android:id="@+id/action_mainFragment_to_encryptedFileActivity"
+            app:destination="@id/encryptedFileActivity" />
+        <action
+            android:id="@+id/action_mainFragment_to_lifecyclesActivity"
+            app:destination="@id/lifecyclesActivity" />
+        <action
+            android:id="@+id/action_mainFragment_to_navigationActivity"
+            app:destination="@id/navigationActivity" />
+    </fragment>
+
+    <activity
+        android:id="@+id/dataBindingActivity"
+        android:name="com.imeepwni.jetpack.function.data_binding.DataBindingActivity"
+        android:label="DataBindingActivity"
+        tools:layout="@layout/activity_data_binding" />
+
+    <activity
+        android:id="@+id/encryptedFileActivity"
+        android:name="com.imeepwni.jetpack.function.encrypted_file.EncryptedFileActivity"
+        android:label="EncryptedFileActivity"
+        tools:layout="@layout/activity_encrypted_file" />
+
+    <activity
+        android:id="@+id/lifecyclesActivity"
+        android:name="com.imeepwni.jetpack.function.lifecycles.LifecyclesActivity"
+        android:label="LifecyclesActivity" />
+    <activity
+        android:id="@+id/navigationActivity"
+        android:name="com.imeepwni.jetpack.function.navigation.NavigationActivity"
+        android:label="activity_navigation"
+        tools:layout="@layout/activity_navigation" />
+</navigation>
